requirements.txt
Flask==2.3.2
Werkzeug==3.1.2
pymongo==4.6.3
dnspython
python-dotenv
gunicorn==21.2.0
